Разделение вьюмодели на две части
